requirements.txt
crewai>=0.80
crewai[tools]
anthropic
python-amazon-sp-api
firecrawl-py
exa-py
sqlalchemy>=2.0
alembic
apscheduler>=3.10
httpx
pydantic>=2.0
pyyaml
python-dotenv
uvicorn
fastapi
pytest
pytest-asyncio
